[jaxrs-js] client can read/write server, more resources needed
This commit enables to read write todos using a singleton.
Before this commit, writing was failing because we were using
a private field.

However, while now we can GET (read) and POST (add element),
we still cannot DELETE or edit a specific todo.

In order to solve the issue, we need to create one resource
(i.e. /1, /2, etc) for each todo, as shown in Vogella tutorial.