[jaxrs-js] view updates properly on every action
Before this commit the view was not updating
with new model info at every action. It was needed
to refresh the page.
This is manly because view API calls where async,
and they were not completed in time before the following
get was called.

After this commit, calls are sync and the view updates properly.

Other improvements:
- responses from actions
- new reference describing what the server should
  respond to calls from the client