[jaxrs-js] clean Model and fix a bug in Model
Identified bug in todoIndex generation.
Before this commit, if adding todo1, todo2, then
removing todo1 and adding a new Todo the addTodo
method in Model was failing due to wrong id generation.

After this commit, a LinkedHashMap hepls us to
keep track of insertions and can be used for Id
generation.

Other activities:
- removed some commented code